license, readme; made script more user-friendly
When there are no paths passed as command-line arguments but
stdin is a terminal, this is likely to be an accidental interactive
use, and a warning is issued.

I've also made this use the system Python 3 (it doesn't use any
recent language features; this may still not work on systems where
python3 is simply called python but that didn't work before
either).

Also, added a type annotation to mypy checks main(), and a
.gitignore line for the mypy cache.